Update README.md
A website which is used to short the link as well as you can give another name to that link also. I used to create this website by the help of NodeJS Express Cors as a backend Server and React as a Frontend server and Chakra as a CSS library for login and signup I used reqres api.